Latest version with comments, plots and variables.
The 'preliminary cleaning' need some help since the second operation isn't working as it should. Added the comments to the initial plots (we could use the 2nd word cloud as the cover image of the report). The comments can be used for the report and make it easier to understand the code and the data. There are a few more variables in the dataset, mainly rates that compare values to some defined standards. (in the notebook there is a precise description for each one of them).In the last section 'Dataset Analysis after cleaning and transforming' there is a very detailed description of all the variables that we have after cleaning and transforming the dataset. This could be very useful for the report. In the last section, a comparison of the df.describe that made at the beginning with the data that we have at the end of the cleaning and transforming. What has to be done is: finish the comparison, create some plots to describe the variables, and study the correlation in order to drop those variables with almost a perfect correlation. 